More localization text.

+++ Src/SCPCB/Src/Config/Difficulty.cpp
@@ -8,7 +8,7 @@ Difficulty::Difficulty() {
     list.push_back(this);
 }
 Difficulty::~Difficulty() {
-    for (int i = 0; i < list.size(); i++) {
+    for (int i = 0; i < (int)list.size(); i++) {
         if (list[i] == this) {
             list.erase(list.begin() + i);
             break;
